+++ database/factories/UserFactory.php
@@ -26,6 +26,8 @@ class UserFactory extends Factory
             'password' => static::$password ??= Hash::make('password'),
             'remember_token' => Str::random(10),
             'role' => 'admin',
+            'locale' => 'en',
+            'timezone' => 'Europe/Budapest',
         ];
     }